Revert "refactor(balancer) major refactor of balancer and lua-resty-dns-client usage in Kong"
This reverts commit 6a8f3e267a2b7bd64d00cef78d55bd55f0088036.
