+++ kong-2.4.1-0.rockspec
@@ -143,12 +143,7 @@ build = {
     ["kong.runloop.handler"] = "kong/runloop/handler.lua",
     ["kong.runloop.certificate"] = "kong/runloop/certificate.lua",
     ["kong.runloop.plugins_iterator"] = "kong/runloop/plugins_iterator.lua",
-    ["kong.runloop.balancer"] = "kong/runloop/balancer/init.lua",
-    ["kong.runloop.balancer.balancers"] = "kong/runloop/balancer/balancers.lua",
-    ["kong.runloop.balancer.healthcheckers"] = "kong/runloop/balancer/healthcheckers.lua",
-    ["kong.runloop.balancer.round_robin"] = "kong/runloop/balancer/round_robin.lua",
-    ["kong.runloop.balancer.targets"] = "kong/runloop/balancer/targets.lua",
-    ["kong.runloop.balancer.upstreams"] = "kong/runloop/balancer/upstreams.lua",
+    ["kong.runloop.balancer"] = "kong/runloop/balancer.lua",
     ["kong.runloop.plugin_servers"] = "kong/runloop/plugin_servers/init.lua",
     ["kong.runloop.plugin_servers.process"] = "kong/runloop/plugin_servers/process.lua",
     ["kong.runloop.plugin_servers.mp_rpc"] = "kong/runloop/plugin_servers/mp_rpc.lua",
