+++ spec/02-integration/05-proxy/10-balancer/04-round-robin_spec.lua
@@ -429,3 +429,4 @@ for _, consistency in ipairs(bu.consistencies) do
 
   end
 end
+
